Implemented authentification

+++ backend/controllers/HotelController.js
@@ -2,23 +2,31 @@ import Hotel from "../models/Hotel.js";
 
 const HotelController = {
   get: async (req, res) => {
-    res.status(200).json(await Hotel.find(req.query).limit(10).skip(0));
+    res.status(200).json(await Hotel.find().limit(10).skip(0));
   },
   add: async (req, res) => {
-    res.status(200).json(await Hotel.create(req.body));
+    res.status(200).json(await Hotel.create({
+      name:req.body.name,
+      user_id: res.locals.user._id
+    }));
   },
   edit: async (req, res) => {
     const { _id, name, description } = req.body;
-    res.status(200).json(
-      await Hotel.findByIdAndUpdate(_id, {
-        name: name,
-        description: description,
-      })
-    );
+    if(res.locals.user._id === _id) {
+      res.status(200).json(
+        await Hotel.findByIdAndUpdate(_id, {
+          name: name,
+          description: description,
+        })
+      );
+    }
+    res.status(401).json("Not Authorized");
   },
   delete: async (req, res) => {
-    console.log(req.query);
-    res.status(200).json(await Hotel.findByIdAndDelete(req.query._id));
+    if(res.locals.user._id === req.query._id){
+      res.status(200).json(await Hotel.findByIdAndDelete(req.query._id));
+    }
+    res.status(401).json("Not Authorized");
   },
 };
 export default HotelController;
